requirements.txt
exceptiongroup==1.3.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-order==1.3.0
tomli==2.4.0
typing_extensions==4.15.0
